# [SPARK-26780][CORE]Improve shuffle read using ReadAheadInputStream 