@10110346 2019-03-08
> 
> 
> Could you please share the performance difference running TPCDS with and without this improvement?

Ok, thanks, I will do it