@kiszk 2019-03-19
Does it make sense to see the total execution time? Is it better to compare execution times for each query?