@10110346 2019-03-25
> 
> 
> BTW, Seems the gain of this change is not obvious...

Yeah, it is not obvious, but the changes are relatively simple.