@10110346 2019-03-25
> 
> 
> Does it make sense to see the total execution time? Is it better to compare execution times for each query?

This change is mainly effective for shuffle reading,  so I don't think there's any essential difference between queries. thanks