@10110346 2019-03-25
@juliuszsompolski  @sitalkedia  You know best about `ReadAheadInputStream`,  do you think it's better to use `ReadAheadInputStream` in this place? thanks.