@10110346 2019-05-07
> 
> 
> Glad to see someone working on this, I had started looking at this a long time back, but I was looking at using the linux readahead call. I assume you didn't try the linux readahead vs this to see performance difference? I know the ReadAheadInputStream was already here so easier to use, but I would also think the os one might be more efficient if it doesn't have to copy into user space and then you copy it again to send it out, the os one just pulls into the page cache.
> 
> In your performance tests, how did you test? if the data was in the page cache then you wouldn't see much benefit, its when the data actually has to be pulled from disk and the readahead does enough to get it into memory before you need it.
> Have you run this through any real spark applications or on a real cluster to see results?

Yeah, I  run this  through many real spark applications  on a real cluster ,  and a little performance improvement, thanks