@tgravescs 2019-05-07
In your performance tests, how did you test? if the data was in the page cache then you wouldn't see much benefit, its when the data actually has to be pulled from disk and the readahead does enough to get it into memory before you need it.